src/python/requirements.txt
aiokafka==0.6.0
faust[fast]==1.10.4
kafka-python==1.4.7
robinhood-aiokafka==1.1.6
aiohttp==3.6.2
confluent-kafka==1.5.0